eval: Break apart tests
Factors out a common test_domain mod and uses it
for various LTL tests.
