+++ lib/bombadil-ltl/src/lib.rs
@@ -5,6 +5,10 @@ pub mod syntax;
 pub mod violation;
 
 #[cfg(test)]
+mod ltl_depth_tests;
+#[cfg(test)]
 mod ltl_equivalences;
 #[cfg(test)]
 mod ltl_snapshot_tests;
+#[cfg(test)]
+mod test_domain;
